Fix Player Position Bug
Fixed the bug where the player appeared to be in the wrong position.

The world was being rendered at (0,0) without the camera offset being applied.

A new property for camera has been added called Offset which can be used to get how far an object should be offset from its original position.  The value from this property is used for the position of the world images.

The player rendering has been moved to the player class and is called by the Tilemap draw method rather than being drawn directly by the Tilemap.