Adjust Jump Physics
Adjusted the jump physics to feel less floaty and allow variation in how high the player can jump.  Pressing the jump button down for longer results in a longer jump.

Added the property "Falling" to Player class which returns True if the player is falling.

Gravity can now only accelerate the player up until a terminal velocity, but other forces can accelerate the player faster.  Gravity is limited to a velocity of -1500px/second as of right now, but this is a temporary value.

In constants, PHYS_GRAVITY and PHYS_JUMP_POWER have been adjusted to work with the aforementioned changes.