Render at smaller size
The camera now has a Surface property to get the surface it renders the game to.

Settings for RENDER_WIDTH, RENDER_HEIGHT and RENDER_SCALE have been added to constants.py to adjust rendering.

Camera now has a new draw method which a target surface to draw the render to is passed as the only argument.  The camera's render is automatically scaled to the size of the target surface.

Gravity and Jump Power have been adjusted to feel more bouncy.

Event handling in main has been slightly reordered.